Move verifyProxyAndImplementation to helper.ts
